:bug: Resolve custom encryption related bugs
