+++ composeApp/src/commonMain/kotlin/com/crosspaste/net/plugin/ServerDecryptionPluginFactory.kt
@@ -25,6 +25,7 @@ class ServerDecryptionPluginFactory(private val secureStore: SecureStore) {
                     headers["secure"]?.let {
                         logger.debug { "server decrypt $appInstanceId" }
                         return@on application.writer {
+                            logger.info { "xxxxxxxxxxxxx" }
                             val processor = secureStore.getMessageProcessor(appInstanceId)
                             val encryptedContent = body.readRemaining().readByteArray()
                             val decrypted = processor.decrypt(encryptedContent)
