+++ composeApp/src/desktopTest/kotlin/com/crosspaste/net/TestServerModule.kt
@@ -1,8 +1,11 @@
 package com.crosspaste.net
 
+import com.crosspaste.app.AppInfo
 import com.crosspaste.app.AppTokenApi
+import com.crosspaste.app.EndpointInfoFactory
 import com.crosspaste.net.plugin.ServerDecryptionPluginFactory
 import com.crosspaste.net.plugin.ServerEncryptPluginFactory
+import com.crosspaste.net.routing.SyncRoutingApi
 import com.crosspaste.net.routing.syncRouting
 import com.crosspaste.secure.SecureKeyPairSerializer
 import com.crosspaste.secure.SecureStore
@@ -13,11 +16,14 @@ import io.ktor.server.plugins.contentnegotiation.*
 import io.ktor.server.routing.*
 
 class TestServerModule(
+    private val appInfo: AppInfo,
     private val appTokenApi: AppTokenApi,
+    private val endpointInfoFactory: EndpointInfoFactory,
     private val secureKeyPairSerializer: SecureKeyPairSerializer,
     private val secureStore: SecureStore,
     private val serverEncryptPluginFactory: ServerEncryptPluginFactory,
     private val serverDecryptionPluginFactory: ServerDecryptionPluginFactory,
+    private val syncRoutingApi: SyncRoutingApi,
 ) : ServerModule {
     override fun installModules(): Application.() -> Unit =
         {
@@ -28,9 +34,12 @@ class TestServerModule(
             install(serverDecryptionPluginFactory.createPlugin())
             routing {
                 syncRouting(
+                    appInfo,
                     appTokenApi,
+                    endpointInfoFactory,
                     secureKeyPairSerializer,
                     secureStore,
+                    syncRoutingApi,
                 )
             }
         }
